training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
onnx>=1.15.0
onnxscript>=0.1.0
onnxruntime>=1.16.0
tqdm>=4.65.0
matplotlib>=3.7.0
